perf(thermostat-led-display): replace many ints with shorts

+++ THERMOSTAT/thermostat-led-display/led-device.h
@@ -6,8 +6,8 @@ class HeatingLedDevice {
 
 private:
   bool _isOn = false;
-  const unsigned int _PIN = settings.HEATING_LED_PIN;
-  const unsigned int _BRIGHTNESS_VALUE = settings.HEATING_LED_BRIGHTNESS_VALUE;
+  const unsigned short _PIN = settings.HEATING_LED_PIN;
+  const unsigned short _BRIGHTNESS_VALUE = settings.HEATING_LED_BRIGHTNESS_VALUE;
 
 public:
   void setup();
@@ -23,8 +23,8 @@ class ErrorLedDevice {
 
 private:
   bool _isOn = false;
-  const unsigned int _PIN = settings.ERROR_LED_PIN;
-  const unsigned int _BRIGHTNESS_VALUE = settings.ERROR_LED_BRIGHTNESS_VALUE;
+  const unsigned short _PIN = settings.ERROR_LED_PIN;
+  const unsigned short _BRIGHTNESS_VALUE = settings.ERROR_LED_BRIGHTNESS_VALUE;
 
 public:
   void setup();
@@ -40,8 +40,8 @@ class DomainLedDevice {
 
 private:
   bool _isOn = false;
-  const unsigned int _PIN = settings.DOMAIN_LED_PIN;
-  const unsigned int _BRIGHTNESS_VALUE = settings.DOMAIN_LED_BRIGHTNESS_VALUE;
+  const unsigned short _PIN = settings.DOMAIN_LED_PIN;
+  const unsigned short _BRIGHTNESS_VALUE = settings.DOMAIN_LED_BRIGHTNESS_VALUE;
 
 public:
   void setup();
