+++ THERMOSTAT/thermostat-led-display/list-utils.h
@@ -1,5 +1,5 @@
 class ListUtils {
 public:
   template<typename T>
-  static T getByIndex(std::list<T> l, int i);
+  static T getByIndex(std::list<T> l, short i);
 };
